bug: Fix broken download link
The download link used in the release badge on the main page has a
leading slash which tells mkdocs to anchor the URL against the base
domain name.

This removes the initial slash from the URL, allowing mkdocs to
correctly use the same path as generated via `docs/download.md`

This was tested with the [site_url][1] configuration setting from mkdocs
to ensure that it is resillient to base path changes.

[1]: https://www.mkdocs.org/user-guide/configuration/#site_url